Moving test utility out of mocks to avoid mock conflict
